Fix Gettr for real
I thought that I had confirmed that the "claimable" field worked to
detect account presence, but it seems not. Random usernames also return
claimable as false.

Change the e_string and m_string back to looking at the "success"
element. If the username exists, the "success" name will be followed
by an object starting with "{" containing user information. If the
username does not exist, we will simply see that the value for the
"success" name has a value of false.

Add a pretty_uri while we're here.