Update wmn-data.json
WeTransfer has to possibility to host users subdomains. When hitting a valid one a 200 is returned which includes the email of the user the account belongs to in a JS-variable called __session__. Test it with curl -s 'https://joe.wetransfer.com/' | grep default_recipient_email. An invalid user yields a status code of 302 and a redirect to wetransfer.com.